Strengthen validation, fix ambiguous APIs, enforce type consistency
- Grades.getLast(): throw NoSuchElementException on empty list instead of returning 0
- Notebook: validate weight >= 0, price >= 0, year in [1970, 2030]
- User: add equals/hashCode/toString
- Calculator.circleArea: reject negative radius
- AdvCalculator: return double consistently from all operations
- Book: add null validation for author and title
- AirportRepository: add null-check on airportName parameter
- Rename MAX_RANDOM_VALUE to RANDOM_BOUND for clarity

+++ kodilla-intro/src/main/java/com/kodilla/AdvCalculator.java
@@ -6,15 +6,15 @@ public class AdvCalculator {
         int a = UserDialogs.getValue();
         int b = UserDialogs.getValue();
         return switch (userSelected) {
-            case "ADD" -> a + b;
-            case "SUB" -> a - b;
+            case "ADD" -> (double) a + b;
+            case "SUB" -> (double) a - b;
             case "DIV" -> {
                 if (b == 0) {
                     throw new ArithmeticException("Division by zero is not allowed");
                 }
                 yield (double) a / b;
             }
-            case "MUL" -> a * b;
+            case "MUL" -> (double) a * b;
             default -> throw new IllegalArgumentException("Unknown operation: " + userSelected);
         };
     }
